+++ src/Finder/QueryRegexFinder.php
@@ -6,7 +6,6 @@ namespace WebThumbnailer\Finder;
 
 use WebThumbnailer\Application\ConfigManager;
 use WebThumbnailer\Application\WebAccess\WebAccess;
-use WebThumbnailer\Application\WebAccess\WebAccessCUrl;
 use WebThumbnailer\Application\WebAccess\WebAccessFactory;
 use WebThumbnailer\Exception\BadRulesException;
 use WebThumbnailer\Utils\FinderUtils;
@@ -63,88 +62,20 @@ class QueryRegexFinder extends FinderCommon
      */
     public function find()
     {
-        $thumbnail = $content = null;
-        $callback = $this->webAccess instanceof WebAccessCUrl
-            ? $this->getCurlCallback($content, $thumbnail)
-            : null;
         list($headers, $content) = $this->webAccess->getContent(
             $this->url,
             (int) ConfigManager::get('settings.default.timeout', 30),
-            (int) ConfigManager::get('settings.default.max_img_dl', 16777216),
-            $callback,
-            $content
+            (int) ConfigManager::get('settings.default.max_img_dl', 16777216)
         );
         if (
             empty($content)
             || empty($headers)
-            || (empty($thumbnail) && strpos($headers[0], '200') === false)
+            || strpos((string) $headers[0], '200') === false
         ) {
             return false;
         }
 
-        // With curl, the thumb is extracted during the download
-        if ($this->webAccess instanceof WebAccessCUrl && ! empty($thumbnail)) {
-            return $thumbnail;
-        }
-
         return $this->extractThumbContent($content);
-    }
-
-    /**
-     * Get a callback for curl write function.
-     *
-     * @param string|null $content   A variable reference in which the downloaded content should be stored.
-     * @param string|null $thumbnail A variable reference in which extracted thumb URL should be stored.
-     *
-     * @return callable CURLOPT_WRITEFUNCTION callback
-     */
-    protected function getCurlCallback(?string &$content, ?string &$thumbnail): callable
-    {
-        $isRedirected = false;
-
-        /**
-         * cURL callback function for CURLOPT_WRITEFUNCTION (called during the download).
-         *
-         * While downloading the remote page, we check that the HTTP code is 200 and content type is 'html/text'
-         * Then we extract the title and the charset and stop the download when it's done.
-         *
-         * Note that when using CURLOPT_WRITEFUNCTION, we have to manually handle the content retrieved,
-         * hence the $content reference variable.
-         *
-         * @param resource $ch   cURL resource
-         * @param string   $data chunk of data being downloaded
-         *
-         * @return int|false length of $data or false if we need to stop the download
-         */
-        return function ($ch, $data) use (&$content, &$thumbnail, &$isRedirected) {
-            $content .= $data;
-            $responseCode = curl_getinfo($ch, CURLINFO_RESPONSE_CODE);
-
-            if (!empty($responseCode) && in_array($responseCode, [301, 302])) {
-                $isRedirected = true;
-                return strlen($data);
-            }
-            if (!empty($responseCode) && $responseCode !== 200) {
-                return false;
-            }
-            // After a redirection, the content type will keep the previous request value
-            // until it finds the next content-type header.
-            if (! $isRedirected || strpos(strtolower($data), 'content-type') !== false) {
-                $contentType = curl_getinfo($ch, CURLINFO_CONTENT_TYPE);
-            }
-            if (!empty($contentType) && strpos($contentType, 'text/html') === false) {
-                return false;
-            }
-            if (empty($thumbnail)) {
-                $thumbnail = $this->extractThumbContent($data);
-            }
-            // We got everything we want, stop the download.
-            if (!empty($responseCode) && !empty($contentType) && !empty($thumbnail)) {
-                return false;
-            }
-
-            return strlen($data);
-        };
     }
 
     /**